+++ config/app-web.php
@@ -23,7 +23,6 @@ $config = [
 	],
 	'components' => [
 		'request'      => [
-			// TODO: move generator to console command.
 			'cookieValidationKey' => env('APP_KEY'),
 		],
 		'response' => [
@@ -72,7 +71,7 @@ $config = [
 			'errorAction' => 'site/error',
 		],
 		'mailer'       => [
-			'class'            => 'yii\swiftmailer\Mailer',
+			'class'            => \yii\swiftmailer\Mailer::class,
 			// send all mails to a file by default. You have to set
 			// 'useFileTransport' to false and configure a transport
 			// for the mailer to send real emails.
@@ -82,7 +81,7 @@ $config = [
 			'traceLevel' => YII_DEBUG ? 3 : 0,
 			'targets'    => [
 				[
-					'class'  => 'yii\log\FileTarget',
+					'class'  => \yii\log\FileTarget::class,
 					'levels' => ['error', 'warning'],
 				],
 			],
@@ -92,17 +91,17 @@ $config = [
 ];
 
 if (YII_ENV_DEV) {
-	// configuration adjustments for 'dev' environment
+	// configuration adjustments for 'dev' environment.
 	$config['bootstrap'][]      = 'debug';
 	$config['modules']['debug'] = [
-		'class' => 'yii\debug\Module',
+		'class' => \yii\debug\Module::class,
 		// uncomment the following to add your IP if you are not connecting from localhost.
 		'allowedIPs' => ['*'],
 	];
 
 	$config['bootstrap'][]    = 'gii';
 	$config['modules']['gii'] = [
-		'class' => 'yii\gii\Module',
+		'class' => \yii\gii\Module::class,
 		// uncomment the following to add your IP if you are not connecting from localhost.
 		'allowedIPs' => ['*'],
 	];