+++ database/migrations/templates/_dropTable.php
@@ -7,7 +7,7 @@
 /* @var $foreignKeys array the foreign keys */
 
 echo $this->render('_dropForeignKeys', [
-    'table' => $table,
-    'foreignKeys' => $foreignKeys,
+	'table' => $table,
+	'foreignKeys' => $foreignKeys,
 ]) ?>
 		$this->dropTable('<?= $table ?>');