+++ database/migrations/templates/createTableMigration.php
@@ -7,16 +7,18 @@
 /* @var $className string the new migration class name without namespace */
 /* @var $namespace string the new migration class namespace */
 /* @var $table string the name table */
+/* @var $tableComment string the comment table */
 /* @var $fields array the fields */
 /* @var $foreignKeys array the foreign keys */
 
 echo "<?php\n";
 if (!empty($namespace)) {
-    echo "\nnamespace {$namespace};\n";
+	echo "\nnamespace {$namespace};\n";
 }
 ?>
 
-use app\modules\base\db\Migration;
+use yii\db\Migration;
+use app\traits\migrations\CreateTableOptions;
 
 /**
  * Handles the creation of table `<?= $table ?>`.
@@ -26,10 +28,12 @@ use app\modules\base\db\Migration;
  */
 class <?= $className ?> extends Migration
 {
+	use CreateTableOptions;
+
 	/**
-	 * @inheritdoc
+	 * {@inheritdoc}
 	 */
-	public function up()
+	public function safeUp()
 	{
 <?= $this->render('_createTable', [
 	'table' => $table,
@@ -37,12 +41,19 @@ class <?= $className ?> extends Migration
 	'foreignKeys' => $foreignKeys,
 ])
 ?>
+<?php if (!empty($tableComment)) {
+	echo $this->render('_addComments', [
+		'table' => $table,
+		'tableComment' => $tableComment,
+	]);
+}
+?>
 	}
 
 	/**
-	 * @inheritdoc
+	 * {@inheritdoc}
 	 */
-	public function down()
+	public function safeDown()
 	{
 <?= $this->render('_dropTable', [
 	'table' => $table,